Uploaded the Birdie, and Made Imp and Eyedevil animation in Animator.
Details for where they are in the folder management.
Assets > sprites > Monsters > Eyedevil for Eyedevil
Assets > sprites > Monsters > Imp for Imp

THe sprites of the birdie animation is out of the folder Birdie. But i've already made the animations within the folder. Don't know how that will effect. So I didn't clean up.